changed testing files
